Reuse guided permissions from menu
